extract AgentLike and rename to LangChainAgent
