+++ backend/src/langchain/utils.ts
@@ -1,84 +1,10 @@
 import { randomUUID } from "crypto";
-import { BaseCallbackHandler } from "@langchain/core/callbacks/base";
-import { CallbackManager } from "@langchain/core/callbacks/manager";
-import { AIMessage, BaseMessage, ReactAgent, ToolMessage } from "langchain";
+import { AIMessage, BaseMessage, ToolMessage } from "langchain";
 import {
-  Agent,
-  AgentMessage,
   AgentTraceMessage,
   AgentTraceMessageType,
   ToolExecution,
 } from "../ports/agent-types";
-
-export class AgentLike<TContext extends Record<string, unknown>>
-  implements Agent<TContext>
-{
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  constructor(private agent: ReactAgent<any>) {}
-
-  async invoke(
-    state: { messages: readonly AgentMessage[] },
-    config: { context: TContext },
-  ) {
-    const agentTrace: AgentTraceMessage[] = [];
-    const pendingToolInputs = new Map<string, string>();
-    const toolExecutions: ToolExecution[] = [];
-    const callbackManager = new CallbackManager();
-
-    callbackManager.addHandler(
-      BaseCallbackHandler.fromMethods({
-        handleLLMEnd(llmResult) {
-          for (const promptGenerations of llmResult.generations) {
-            for (const generation of promptGenerations) {
-              if (!("message" in generation)) {
-                // Skip if the generation does not contain a message
-                continue;
-              }
-
-              const message = generation.message;
-
-              if (!(message instanceof AIMessage)) {
-                // Skip to narrow down to AIMessage instances
-                continue;
-              }
-
-              const agentTraceMessages = extractAgentTraceTexts(message);
-              agentTrace.push(...agentTraceMessages);
-            }
-          }
-        },
-        handleToolStart(_tool, input, runId) {
-          pendingToolInputs.set(runId, input);
-        },
-        handleToolEnd(output, runId) {
-          const agentTraceMessage = extractAgentTraceToolResult(output);
-          agentTrace.push(agentTraceMessage);
-
-          toolExecutions.push({
-            tool: output.name || "Unknown tool",
-            input: pendingToolInputs.get(runId) ?? "Unknown arguments",
-            output: output.content ? String(output.content) : "Unknown result",
-          });
-          pendingToolInputs.delete(runId);
-        },
-      }),
-    );
-
-    const messages = state.messages.map((message) => ({
-      role: message.role,
-      content: message.content,
-    }));
-
-    const response = await this.agent.invoke(
-      { ...state, messages },
-      { ...config, callbacks: callbackManager },
-    );
-
-    const answer = extractLastMessageText(response.messages)?.trim();
-
-    return { answer, agentTrace, toolExecutions };
-  }
-}
 
 export function extractLastMessageText(
   messages: BaseMessage[],
@@ -155,7 +81,9 @@ export function extractToolExecutions(
   return Array.from(toolExecutionsMap.values());
 }
 
-function extractAgentTraceTexts(message: AIMessage): AgentTraceMessage[] {
+export function extractAgentTraceTexts(
+  message: AIMessage,
+): AgentTraceMessage[] {
   const result: AgentTraceMessage[] = [];
 
   if (typeof message.content === "string") {
@@ -205,7 +133,9 @@ function extractAgentTraceTexts(message: AIMessage): AgentTraceMessage[] {
   return result;
 }
 
-function extractAgentTraceToolResult(message: ToolMessage): AgentTraceMessage {
+export function extractAgentTraceToolResult(
+  message: ToolMessage,
+): AgentTraceMessage {
   const toolName = message.name || "Unknown tool";
   const rawToolOutput = message.content;
   let toolOutput: string;
